% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkg2-dependencies.R
\name{add_mypkg2_deps}
\alias{add_mypkg2_deps}
\title{mypkg2 dependencies utils}
\usage{
add_mypkg2_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches mypkg2. dependencies to the given tag
}
